requirements.txt
alembic==1.7.7
amqp==5.1.0
atomicwrites==1.4.0
attrs==21.2.0
backports.cached-property==1.0.1
cached-property==1.5.2
classic-app-layer==0.0.1
classic-aspects==0.0.1
classic-components==0.0.1
classic-http-api==0.0.2
classic-http-auth==0.2.1
classic-messaging==0.0.2
classic-messaging-kombu==0.0.3
classic-sql-storage==0.0.1
click==7.1.2
colorama==0.4.4
falcon==3.0.1
greenlet==1.1.2
hupper==1.10.3
importlib-metadata==4.11.3
importlib-resources==5.4.0
iniconfig==1.1.1
kombu==5.1.0
Mako==1.2.0
MarkupSafe==2.1.1
marshmallow==3.15.0
mysqlclient==2.1.0
packaging==21.3
pika==0.13.0
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
PyAMQP==0.1.0.7
pydantic==1.8.2
PyJWT==2.0.1
pyparsing==3.0.7
pytest==7.1.0
python-dotenv==0.20.0
python-json-logger==2.0.2
SQLAlchemy==1.4.32
tomli==2.0.1
typing-extensions==4.1.1
vine==5.0.0
waitress==2.1.0
zipp==3.7.0
